refactor(vite-plugin-angular): extract encapsulation plugin into separate file
Move the inline encapsulation plugin, isComponentStyleSheet, and
getComponentStyleSheetMeta into encapsulation-plugin.ts. The plugin
factory takes a shouldExternalizeStyles callback to access the closure
state it needs.
