+++ packages/vite-plugin-angular/src/lib/angular-vite-plugin.ts
@@ -21,7 +21,6 @@ import * as vite from 'vite';
 import * as compilerCli from '@angular/compiler-cli';
 import { createRequire } from 'node:module';
 import * as ts from 'typescript';
-import * as ngCompiler from '@angular/compiler';
 import { globSync } from 'tinyglobby';
 import {
   defaultClientConditions,
@@ -69,6 +68,11 @@ import {
   buildStylePreprocessor,
   validateTailwindConfig,
 } from './tailwind-plugin.js';
+import {
+  encapsulationPlugin,
+  isComponentStyleSheet,
+  getComponentStyleSheetMeta,
+} from './encapsulation-plugin.js';
 import { angularVitestPlugins } from './angular-vitest-plugin.js';
 import {
   createJitResourceTransformer,
@@ -1663,32 +1667,7 @@ export function angular(options?: PluginOptions): Plugin[] {
     routerPlugin(),
     angularFullVersion < 190004 && pendingTasksPlugin(),
     nxFolderPlugin(),
-    // Encapsulation runs in enforce: 'post' so that @tailwindcss/vite
-    // (enforce: 'pre') fully resolves @apply directives — including those
-    // inside :host {} — before Angular's ShadowCss rewrites selectors.
-    // Previously this ran in the main plugin's normal-phase transform,
-    // which could race with @tailwindcss/vite depending on plugin
-    // registration order. (#2293)
-    {
-      name: '@analogjs/vite-plugin-angular:encapsulation',
-      enforce: 'post',
-      transform(code: string, id: string) {
-        if (shouldExternalizeStyles() && isComponentStyleSheet(id)) {
-          const { encapsulation, componentId } = getComponentStyleSheetMeta(id);
-          if (encapsulation === 'emulated' && componentId) {
-            debugStylesV('applying emulated view encapsulation (post)', {
-              stylesheet: id.split('?')[0],
-              componentId,
-            });
-            const encapsulated = ngCompiler.encapsulateStyle(code, componentId);
-            return {
-              code: encapsulated,
-              map: null,
-            };
-          }
-        }
-      },
-    } satisfies Plugin,
+    encapsulationPlugin(shouldExternalizeStyles),
   ].filter(Boolean) as Plugin[];
 
   function findIncludes() {
@@ -2988,28 +2967,6 @@ function markModuleSelfAccepting(mod: ModuleNode): ModuleNode {
     ...mod,
     isSelfAccepting: true,
   } as ModuleNode;
-}
-
-function isComponentStyleSheet(id: string): boolean {
-  return id.includes('ngcomp=');
-}
-
-function getComponentStyleSheetMeta(id: string): {
-  componentId: string;
-  encapsulation: 'emulated' | 'shadow' | 'none';
-} {
-  const params = new URL(id, 'http://localhost').searchParams;
-  const encapsulationMapping = {
-    '0': 'emulated',
-    '2': 'none',
-    '3': 'shadow',
-  };
-  return {
-    componentId: params.get('ngcomp')!,
-    encapsulation: encapsulationMapping[
-      params.get('e') as keyof typeof encapsulationMapping
-    ] as 'emulated' | 'shadow' | 'none',
-  };
 }
 
 /**
